requirements.txt
streamlit
psycopg2-binary
python-dotenv
pandas
plotly
altair